move constants to outlet.model.ts
