--- projects/storefrontlib/src/cms-structure/outlet/outlet-constants.ts
@@ -1,3 +1,0 @@
-export const AVOID_STACKED_OUTLETS = false;
-
-export const USE_STACKED_OUTLETS = true;
