+++ projects/storefrontlib/src/cms-structure/outlet/outlet.model.ts
@@ -3,3 +3,6 @@ export enum OutletPosition {
   BEFORE = 'before',
   AFTER = 'after',
 }
+
+export const AVOID_STACKED_OUTLETS = false;
+export const USE_STACKED_OUTLETS = true;
