+++ projects/storefrontlib/src/cms-structure/outlet/outlet.service.ts
@@ -1,6 +1,5 @@
 import { Injectable, TemplateRef } from '@angular/core';
-import { AVOID_STACKED_OUTLETS } from './outlet-constants';
-import { OutletPosition } from './outlet.model';
+import { AVOID_STACKED_OUTLETS, OutletPosition } from './outlet.model';
 
 @Injectable({
   providedIn: 'root',
